+++ supply-service/src/main/java/com/ozz/atlas/supply/subpurchaseorder/dtos/SubPurchaseOrderItemResponse.java
@@ -24,8 +24,8 @@ public class SubPurchaseOrderItemResponse {
     private String unit;
     private BigDecimal unitPrice;
     private BigDecimal lineAmount;
-    private BigDecimal orderedQty;
-    private BigDecimal confirmedQty;
+    private Long orderedQty;
+    private Long confirmedQty;
     private LocalDate requiredDate;
     private SubPurchaseOrderLineStatus lineStatus;
     private LocalDateTime createdAt;
